Update 02.proof-of-work.ipynb
Really sorry about making a mess. i am a totally freshman of Github and i cannot find the Edit Button just now. now i find it and having a try for next step. Sorry again for making any troubles for you. 
And i think maybe it would be better and more convenient for a freshman to github if there is another tips between step 3 create branch and step 4 commit changes.